兼容性调整

+++ examples/single_chip/camera_web_server/main/app_uart.c
@@ -36,17 +36,16 @@ static const char *TAG = "uart_events";
 #define RD_BUF_SIZE (BUF_SIZE)
 static QueueHandle_t uart0_queue;
 
-static void uart_event_task(void *arg)
-{
+static void uart_event_task(void *arg) {
     uart_event_t event;
 //    size_t buffered_size;
-    uint8_t* dtmp = (uint8_t*) malloc(RD_BUF_SIZE);
-    while(1) {
+    uint8_t *dtmp = (uint8_t *) malloc(RD_BUF_SIZE);
+    while (1) {
         //Waiting for UART event.
-        if(xQueueReceive(uart0_queue, (void * )&event, (portTickType)portMAX_DELAY)) {
+        if (xQueueReceive(uart0_queue, (void *) &event, (portTickType) portMAX_DELAY)) {
             bzero(dtmp, RD_BUF_SIZE);
             //ESP_LOGI(TAG, "uart[%d] event:", EX_UART_NUM);
-            switch(event.type) {
+            switch (event.type) {
                 //Event of UART receving data
                 /*We'd better handler data event fast, there would be much more data events than
                 other types of events. If we take too much time on data event, the queue might
@@ -54,10 +53,10 @@ static void uart_event_task(void *arg)
                 case UART_DATA:
                     //ESP_LOGI(TAG, "[UART DATA]: %d", event.size);
                     uart_read_bytes(EX_UART_NUM, dtmp, event.size, portMAX_DELAY);
-                    ESP_LOGI(TAG, "[DATA EVT]:%d:",sizeof(dtmp));
-                    uart_write_bytes(EX_UART_NUM, (const char*) dtmp, event.size);
+                    ESP_LOGI(TAG, "[DATA EVT]:%u:", (uint32_t) sizeof(dtmp));
+                    uart_write_bytes(EX_UART_NUM, (const char *) dtmp, event.size);
                     break;
-                //Event of HW FIFO overflow detected
+                    //Event of HW FIFO overflow detected
                 case UART_FIFO_OVF:
                     ESP_LOGI(TAG, "hw fifo overflow");
                     // If fifo overflow happened, you should consider adding flow control for your application.
@@ -66,7 +65,7 @@ static void uart_event_task(void *arg)
                     uart_flush_input(EX_UART_NUM);
                     xQueueReset(uart0_queue);
                     break;
-                //Event of UART ring buffer full
+                    //Event of UART ring buffer full
                 case UART_BUFFER_FULL:
                     ESP_LOGI(TAG, "ring buffer full");
                     // If buffer full happened, you should consider encreasing your buffer size
@@ -75,55 +74,54 @@ static void uart_event_task(void *arg)
                     xQueueReset(uart0_queue);
                     break;
 #if 0
-                //Event of UART RX break detected
-                case UART_BREAK:
-                    ESP_LOGI(TAG, "uart rx break");
-                    break;
-                //Event of UART parity check error
-                case UART_PARITY_ERR:
-                    ESP_LOGI(TAG, "uart parity error");
-                    break;
-                //Event of UART frame error
-                case UART_FRAME_ERR:
-                    ESP_LOGI(TAG, "uart frame error");
-                    break;
-                //UART_PATTERN_DET
-                case UART_PATTERN_DET:
-                    uart_get_buffered_data_len(EX_UART_NUM, &buffered_size);
-                    int pos = uart_pattern_pop_pos(EX_UART_NUM);
-                    ESP_LOGI(TAG, "[UART PATTERN DETECTED] pos: %d, buffered size: %d", pos, buffered_size);
-                    if (pos == -1) {
-                        // There used to be a UART_PATTERN_DET event, but the pattern position queue is full so that it can not
-                        // record the position. We should set a larger queue size.
-                        // As an example, we directly flush the rx buffer here.
-                        uart_flush_input(EX_UART_NUM);
-                    } else {
-                        uart_read_bytes(EX_UART_NUM, dtmp, pos, 100 / portTICK_PERIOD_MS);
-                        uint8_t pat[PATTERN_CHR_NUM + 1];
-                        memset(pat, 0, sizeof(pat));
-                        uart_read_bytes(EX_UART_NUM, pat, PATTERN_CHR_NUM, 100 / portTICK_PERIOD_MS);
-                        ESP_LOGI(TAG, "read data: %s", dtmp);
-                        ESP_LOGI(TAG, "read pat : %s", pat);
-                    }
-                    break;
+                    //Event of UART RX break detected
+                    case UART_BREAK:
+                        ESP_LOGI(TAG, "uart rx break");
+                        break;
+                    //Event of UART parity check error
+                    case UART_PARITY_ERR:
+                        ESP_LOGI(TAG, "uart parity error");
+                        break;
+                    //Event of UART frame error
+                    case UART_FRAME_ERR:
+                        ESP_LOGI(TAG, "uart frame error");
+                        break;
+                    //UART_PATTERN_DET
+                    case UART_PATTERN_DET:
+                        uart_get_buffered_data_len(EX_UART_NUM, &buffered_size);
+                        int pos = uart_pattern_pop_pos(EX_UART_NUM);
+                        ESP_LOGI(TAG, "[UART PATTERN DETECTED] pos: %d, buffered size: %d", pos, buffered_size);
+                        if (pos == -1) {
+                            // There used to be a UART_PATTERN_DET event, but the pattern position queue is full so that it can not
+                            // record the position. We should set a larger queue size.
+                            // As an example, we directly flush the rx buffer here.
+                            uart_flush_input(EX_UART_NUM);
+                        } else {
+                            uart_read_bytes(EX_UART_NUM, dtmp, pos, 100 / portTICK_PERIOD_MS);
+                            uint8_t pat[PATTERN_CHR_NUM + 1];
+                            memset(pat, 0, sizeof(pat));
+                            uart_read_bytes(EX_UART_NUM, pat, PATTERN_CHR_NUM, 100 / portTICK_PERIOD_MS);
+                            ESP_LOGI(TAG, "read data: %s", dtmp);
+                            ESP_LOGI(TAG, "read pat : %s", pat);
+                        }
+                        break;
 #endif
-                //Others
+                    //Others
                 default:
                     ESP_LOGI(TAG, "uart event type: %d", event.type);
                     break;
             }
-                //ESP_LOGD(TAG, "uart stack : %d", uxTaskGetStackHighWaterMark(NULL));
+            //ESP_LOGD(TAG, "uart stack : %d", uxTaskGetStackHighWaterMark(NULL));
         }
 
-    //ESP_LOGI(TAG, "%sFree heap: %u",__func__ ,xPortGetFreeHeapSize());
+        //ESP_LOGI(TAG, "%sFree heap: %u",__func__ ,xPortGetFreeHeapSize());
     }
     free(dtmp);
     dtmp = NULL;
     vTaskDelete(NULL);
 }
 
-void uart_init()
-{
+void uart_init() {
     esp_log_level_set(TAG, ESP_LOG_INFO);
 
     /* Configure parameters of an UART driver,
@@ -151,6 +149,5 @@ void uart_init()
     //Create a task to handler UART event from ISR
     xTaskCreate(uart_event_task, "uart_event_task", 2048, NULL, 5, NULL);
 
-    
 
 }
